backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.23
websockets>=12.0
crewai>=0.1.0
openai>=1.3.0
langchain>=0.0.350
asyncio-mqtt>=0.13.0
redis>=5.0.1
celery>=5.3.4
httpx>=0.25.2
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1